refactor(cli): use the generated OpenAPI types, drop engine/api
The hand-written engine/api/types.go duplicated what oapi-codegen already
emits into internal/api from openapi.gen.json, so the Go seam was not
actually machine-checked. The codegen config now uses initialism naming
(ID, SessionID) and non-pointer optionals so the generated shapes match
what the CLI used; callers cast at the few enum and int-width seams.
The whoami org gains a named OrgRef schema and PushMeta no longer leaks
the Session user field.

+++ cli/cmd_list.go
@@ -42,7 +42,7 @@ func cmdList(args []string) error {
 	for _, s := range sessions {
 		fmt.Fprintf(w, "%s\t%s\t%s\t%s\t%.60s\t%d\t%s\t%s\n",
 			s.ID[:8], s.Agent, s.Host, oneOf(s.Project, "-"), oneOf(s.Title, "-"),
-			s.Files, humanSize(s.Size), humanTime(s.UpdatedAt))
+			s.Files, humanSize(int64(s.Size)), humanTime(s.UpdatedAt))
 	}
 	return w.Flush()
 }
